+++ lib/src/global/ui/widgets/custom/custom_widgets_barrel.dart
@@ -1,3 +1,4 @@
+export 'package:nobook/src/global/ui/widgets/custom/scrollbar/custom_scrollbar.dart';
 export 'package:nobook/src/global/ui/widgets/custom/value_listenables/change_notifier_builder.dart';
 export 'package:nobook/src/global/ui/widgets/custom/value_listenables/listenable_builder.dart';
 export 'package:nobook/src/global/ui/widgets/custom/value_listenables/value_listenable_selector.dart';
